requirements.txt
pyioc3
dnry-config
dnry-srvhost-builder
pyyaml
